separate delay functions from timeout functions
